requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=- requirements.in
#
alembic==1.9.2
    # via
    #   flask-alembic
    #   flask-migrate
anyio==3.6.2
    # via starlette
ariadne==0.17.1
    # via -r requirements.in
attrs==22.2.0
    # via
    #   -r requirements.in
    #   cattrs
    #   pytest
black==23.1.0
    # via -r requirements.in
cattrs==22.2.0
    # via -r requirements.in
click==8.1.3
    # via
    #   black
    #   flask
coverage[toml]==7.1.0
    # via pytest-cov
exceptiongroup==1.1.0
    # via
    #   cattrs
    #   pytest
flake8==6.0.0
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-alembic
    #   flask-cors
    #   flask-marshmallow
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-alembic==2.0.1
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask-marshmallow==0.14.0
    # via -r requirements.in
flask-migrate==4.0.4
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via
    #   -r requirements.in
    #   flask-alembic
    #   flask-migrate
flask-wtf==1.1.1
    # via -r requirements.in
graphql-core==3.2.3
    # via ariadne
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via anyio
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==3.19.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.28.1
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==23.0
    # via
    #   black
    #   marshmallow
    #   marshmallow-sqlalchemy
    #   pytest
pathspec==0.11.0
    # via black
platformdirs==2.6.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest-html
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytest==7.2.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
pytest-cov==4.0.0
    # via pytest-cover
pytest-cover==3.0.0
    # via pytest-coverage
pytest-coverage==0.0
    # via -r requirements.in
pytest-html==3.2.0
    # via -r requirements.in
pytest-metadata==2.0.4
    # via pytest-html
python-json-logger==2.0.4
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-cors
    #   flask-marshmallow
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.1
    # via
    #   alembic
    #   flask-alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
starlette==0.23.1
    # via ariadne
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.4.0
    # via
    #   ariadne
    #   sqlalchemy
werkzeug==2.2.2
    # via flask
wtforms==3.0.1
    # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
